Face Matching Web Application/requirements.txt
mtcnn==0.1.0
tensorflow==2.3.1
keras==2.4.3
keras-vggface==0.6
keras_applications==1.0.8
PyYAML
tqdm
scikit-learn
streamlit
bing-image-downloader

#local packages-
-e .